More refactoring--adjust ErrorObj so that it's functionality is all inline (see below).
ErrorObj is a class that used to be in the MessageLogger subdirectory.
To reduce the complexity of the system, it was moved to the Utilities
subdirectory.  However, this placed an undue burden on users,
requiring them to link to the MF_Utilities library for all code that
uses the message facility.  To get around this problem, the
functionality of the ErrorObj insertion operators was placed entirely
in the header file, with suitable inlining.

This will increase the library sizes for anything that #include's
messagefacility/MessageLogger/MessageLogger.h.  In practice, however,
I expect (hope?) that the sizes will not increase substantially since
the only insertion operator defined in the ErrorObj.cc file was that
for string literals (of types char const* and char const* const).  To
avoid a ballooning of library sizes due to inlining of:

  ErrorObj& operator<<(ErrorObj&, const char[]) {...}

that signature has been disabled by using suitable invocations of
std::enable_if and requiring an implicit conversion from const char*
to std::string.