Use MF_PLUGIN_PATH env var for plugins is present
As discussed in https://cdcvs.fnal.gov/redmine/issues/17900,
MessageFacility plugins do not work on macOS without SIP being
disabled. The two BasicPluginFactory instances of MessageLoggerScribe
default to using the system dynamic loader environment variable,
which SIP does not pass on to new processes.

Per discussion in #17900, construct the BasicPluginFactory instances
with a search path chosen at runtime so that:

- The environment variable `MF_PLUGIN_PATH` is used if set
- The default OS lib path is used otherwise

At present, this is tested manually on mac OS and confirmed to work.
Additional testing is needed on Linux/macOS to confirm both use
cases.